requirements.txt
Django==2.0.5
django-braces==1.13.0
django-model-utils==3.2.0
django-star-ratings==0.8.0
django-tinymce4-lite==1.7.5
jsmin==2.2.2
python-slugify==3.0.3
pytz==2019.2
swapper==1.1.1
text-unidecode==1.2
